tr: move TapTreeIter into taptree module
We are going to move TapTree into the module (hence the name), and
before we can do this we need to move the iterator.
